refactor: Extract SearchHeader from SearchPage
